+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* Fixed #124 remove reuse_address=True from async server/client
+
 * Fixed #121 Unknown attribute key error
 
 2.2 - Oct 19, 2019